Re-enable the StorageRewrite pass
The StorageRewrite pass was disabled because of it's iteraction with
the TI DMA transformation pass. The rewrite pass flattened dimensions
associated with Allocate statements. This flattening causes problems for
the C7xDMAPass. Updated the StorageRewrite to retain the original
dimensions if there is no merging of allocations.

Enabling the StorageRewrite pass hoists allocations out of loops where
possible. This behavior is leveraged by the C7x code generator when it
converts Allocate statements to L2 allocations.