tvmc: prefer the C++ generated option --target-tidl-option
We initially added options in python side which is --tidl-option. But
the consistent way of adding target specific options in tvmc is to add
them in the C++ side. We already have added these, so each option
appeared twice in the options list, one as --tidl-option and another
as --target-tidl-option. This commit removes the --tidl-option and
favour the one defined in C++.